Remove tags route handler
- Fetch actual tag details for TagDetails
- Disable `generateStaticParams` for TagDetails to ensure SSR
